requirements.txt
astroid==2.4.2
cachetools==4.1.0
certifi==2020.4.5.2
cffi==1.14.0
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cryptography==2.9.2
decorator==4.4.2
dialogflow==1.0.0
feedparser==5.2.1
Flask==1.1.2
future==0.18.2
fuzzywuzzy==0.18.0
gnewsclient==1.12
google-api-core==1.20.0
google-auth==1.17.2
googleapis-common-protos==1.52.0
grpcio==1.29.0
gunicorn==20.0.4
idna==2.9
importlib==1.0.4
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
portalocker==1.7.0
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pylint==2.5.3
python-telegram-bot==12.7
pytz==2020.1
pywin32==228
requests==2.23.0
rsa==4.6
six==1.15.0
toml==0.10.1
tornado==6.0.4
urllib3==1.25.9
Werkzeug==1.0.1
wrapt==1.12.1
